Implement std::convert::From for CommandArgument
We can create CommandArgument from a list of Strings.
build_command_map did exactly that. Now just use it as the
from() function for std::convert::From, and get some
flexibility for free.

CommandArgument was an alias for an external type, and From is
an external trait. But only traits from this crate can be implemented
for external types. So use the new type pattern to turn CommandArgument
into a type from this crate.

Note that the conversion is still silently lossy, which is not ideal.